requirements.txt
amqp==2.3.2
billiard==3.5.0.5
cairocffi==0.9.0
CairoSVG==2.2.1
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.1.4
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.9.0
html5lib==1.0.1
idna==2.8
kombu==4.2.2
Markdown==3.0.1
Pillow==5.3.0
pycparser==2.19
pydash==4.7.4
Pyphen==0.9.5
pytz==2018.7
redis==3.0.1
requests==2.21.0
six==1.12.0
tinycss2==0.6.1
urllib3==1.24.1
vine==1.1.4
WeasyPrint==43
webencodings==0.5.1
